\docType{package}
\name{Rngq-package}
\alias{Rngq}
\alias{Rngq-package}
\title{Process NGQ peptide-level data}
\description{
  Data processing functions for Next Generation
  Quantification data
}
\details{
  \tabular{ll}{ Package: \tab Rngq\cr Type: \tab Package\cr
  Version: \tab 0.1\cr Date: \tab 2012-04-27\cr License:
  \tab MIT\cr Depends: \tab XLConnect\cr }

  This package provides a few function to process the
  peptide-level quantifications extracted by NGQ.
}
\author{
  Marc Kirchner
}
\references{
  \url{http://massquito.net}
}
\keyword{package}

