\docType{data}
\name{xilac_peptides}
\alias{xilac_peptides}
\title{A few peptide quantifications from an early XILAC dataset.}
\description{
  A few peptide quantifications from an early XILAC
  dataset.
}
\author{
  Marc Kirchner
}
\references{
  \url{steenlab.org}
}
\keyword{data}
\keyword{datasets}

